+++ packages/winnow/src/options/index.js
@@ -12,8 +12,7 @@ const {
   normalizeLimit,
   normalizeGeometry,
   normalizeOffset,
-  normalizeProjection,
-  normalizeInSR
+  normalizeProjection
 } = require('./normalizeOptions')
 const { normalizeClassification } = require('./normalizeClassification')
 
@@ -21,7 +20,6 @@ function prepare (options, features) {
   const prepared = _.merge({}, options, {
     collection: normalizeCollection(options, features),
     where: normalizeWhere(options),
-    inSR: normalizeInSR(options),
     geometry: normalizeGeometry(options),
     spatialPredicate: normalizeSpatialPredicate(options),
     fields: normalizeFields(options),